+++ include/mandelbrot.h
@@ -1,6 +1,6 @@
 #ifndef MANDELBROT_CALC
 #define MANDELBROT_CALC
 
-int calculateMandelbrot(float x0, float y0, int iterations);
+int calculateMandelbrot(double x0, double y0, int iterations);
 
 #endif
